test: improve stress test
